Created Makefiel

+++ libft/ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: jschneid <jschneid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/07 08:08:57 by jschneid          #+#    #+#             */
-/*   Updated: 2022/08/18 15:55:44 by jschneid         ###   ########.fr       */
+/*   Updated: 2022/05/26 09:26:43 by jschneid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	*ft_calloc(size_t count, size_t size)
 {
 	void	*rtn;
 
-	if (size > 0 && (INT_MAX / size) < count)
+	if (size > 0 && (SIZE_MAX / size) < count)
 		return (NULL);
 	rtn = malloc(count * size);
 	if (rtn == NULL)
